+++ Magic/CHANGELOG.md
@@ -23,6 +23,7 @@
  - Fix overlapping casts overwriting each others' variables
  - Fix blocked-spells and allowed-spells flags (broken in 7.9.2)
  - The SpawnEntity action now forces spawned mobs to ignore their owner by default. Set "owned: false" to prevent this.
+ - Variables can now be pre-defined in a spell, in case you want to use them in base spell parameters
  
 # 7.9.2